refactor(settings): modularize material form typeahead widgets
